feats done

+++ app/models/answer.py
@@ -5,7 +5,7 @@ class Answer(db.Model):
     __tablename__ = 'answers'
     id = db.Column(db.Integer, primary_key=True)
     answer = db.Column(db.Text)
-    rating = db.Column(db.Integer)
+    rating = db.Column(db.Float)
     user_id = db.Column(db.Integer, db.ForeignKey('users.id'))
     question_id = db.Column(db.Integer, db.ForeignKey('questions.id'))
     club_id = db.Column(db.Integer, db.ForeignKey('clubs.id'))
